Rename generic to abstract
