+++ pysqa/wrapper/abstract.py
